requirements.txt
argcomplete==1.8.2
asn1crypto==0.24.0
backcall==0.1.0
beautifulsoup4==4.5.3
bleach==3.0.0
cached-property==1.5.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
decorator==4.3.0
defusedxml==0.5.0
dnspython==1.15.0
docx2txt==0.6
EbookLib==0.15
entrypoints==0.2.3
exchangelib==1.12.0
fpdf==1.7.2
future==0.16.0
idna==2.7
ipykernel==5.0.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
isodate==0.6.0
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
lxml==4.2.5
mailer==0.8.1
MarkupSafe==1.0
mistune==0.8.3
nb-conda==2.2.1
nb-conda-kernels==2.1.1
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.7.0
ntlm-auth==1.2.0
numpy==1.15.2
opencv-python==3.4.3.18
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
pika==0.12.0
Pillow==5.3.0
pipp==0.0.1
pocketsphinx==0.1.3
prometheus-client==0.4.0
prompt-toolkit==2.0.5
ptyprocess==0.6.0
pycparser==2.19
pycryptodome==3.6.6
Pygments==2.2.0
pyinotify==0.9.6
PyPDF2==1.26.0
pytesseract==0.2.5
python-dateutil==2.7.3
python-dotenv==0.9.1
python-pptx==0.6.5
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
requests==2.19.1
requests-ntlm==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.10.0
SpeechRecognition==3.6.3
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
tzlocal==1.5.1
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
xlrd==1.0.0
XlsxWriter==1.1.1
